DevicePolicy: Fix lockNow() failing to turn off screen
Pawit Pornkitprasan's patch from ICS.  This issue affects JB as it did ICS.
Original patch: http://review.cyanogenmod.com/#/c/14845/

---
A fix introduced in 93c518e4f8abd98f87cda1712b30a5a86cfa60dd
causes lockNow() to sometimes (most of the time) fail to turn
off the screen. The screen will turn off, but then immediately turn
on showing the lock screen.

This is caused by getWindowManager().lockNow() call which calls the
method to lock the screen, which the PowerManager service also
call in parallel, leading to a race condition.

The fix is to only call getWindowManager().lockNow() if the screen
is not on.

Change-Id: Ic90d41bd9f52e3ad0924d3a096162e5083c4d4c7