Frameworks: Lazy dev brings white back to battery styles (1/2)
Bug: Battery color white did not stick all other did. Problem is that
ffffffff convert to int is a to big value for integer. Nice solution use
long all the time....but this would be much changes I am toooo lazy for.
Due that ffffffff is the only value out of range it is set to -1, and -2 is
theme default color.

cheers

the lazy dev

Change-Id: If8c3511f39fb6e9d34f152b300a036fb54dfd204