Correct MediaAudioTrackTest testPlaybackHeadPositionAfterStop
When calling AudioTrack.stop the system needs some time to stop
the audiotrack and set the position to 0.
The current sleep time of 100ms is not enough.
Update the test to use the same strategy as the CTS test,
which is to call getPlaybackHeadPosition multiple time before
giving up and failing the test.

Change-Id: Ie938929c009c41e101782b4a46ac01a0aa73ee34