Frameworks: do not depend on quick_settings_num_columns value
the overlay handling from google does not really make sense
and brought us already on 4.2 on weird behaviours sometimes.
Due that we have included tiles per row and duplicate feature
it makes sens to only depend on the values which are programmatically
set. As well I did not find one device or theme which manipulates
the mentioned values.

So let us do it all programmatically and the weird behaviours are
a problem of the past

cheers

Change-Id: I4ce7c5e9c30377983e61880315f60bfbd886a8a4