fb: allow to disable quicksettings
on 4.3 we had an extra settings option for it. Due that we want
to reduce the settings to improve the UX and due that
we are doing it this way on all dslv settings we change
the old 4.3 way.

Now it is very easy like dslv.....remove all tiles and QS
is completly disabled and cannot be opend any more. Notification
drawer shows the settings icon on the top right in this case.
As well compared to 4.3 this works on phones and tablets

Change-Id: I4cebc43e41cb5f54d17a4e8b3c94bb2a79d0879c
TODO: add a hint for the user in settings when disabled like in dslv