QS: fix wrong measurement on 4 or 5 tiles per row on settingspanelview
Issue was only visible on devices which have settingspanelview (tablets - not flippable)
on boot up on landscape and had 4 or 5 tiles per row set. Problem is that due we have
the same layout situation like in portrait on this devices that the layoutparameter
height needs to be calculated in landscape as well.

Solve it with forcing (on only settingspanelview devices) the calculation on landscape
as well

Change-Id: Ib87485e0892abe798006ba5ffaaf5061d560afd6