SysUI: Don't recreate controllers.
The various controllers, such as mLocationController, get recreated
every time the status bar is recreated due to TRDS.  We
should only need to create these if they are null, otherwise use
the controller as is.  Also, each of these controlers registers
themselves as a broadcast receiver and are never unregistered before
recreating them.

Change-Id: I6215aae5a82b7ced8f0f25cf592a804d38825260