FB: Add color values for ease of TRDS theming
This adds color values that I can use for TRDS theming.  I needed values
for colors that are white when TRDS is on and black when off, and vice
versa. Google is implementing framework public value redirects in the
smali code of several Google apps I'm trying to bring TRDS compatibility
to. This will make it much easier to redirect text colors.

PS6: Remove whitespaces. Sorry 'bout that!
PS7: Remove unnecessary duplicated colors in holodark.
Change-Id: Ia3f21b6891ca59d98b79d3ca6e32f6710d9321f3