Frameworks: Actually start to use HEADS_UP_REQUESTED extra
Prepairing slim heads up feature, we do this here to allow apps to
send a heads up notification no matter what.

Google introduced this flag but is not using it in the actual API

Take HEADS_UP_REQUESTED into account to indicate that this notification
is a good heads up candidate.

How to use it you can see in email or mms app

Change-Id: I5404db7a8efb7641188d66998481caa45d5cd661