Properly clean up when setting new InputFilter.
The single InputFilter that is set in InputManagerService was replaced
with a chain of filters that are all listening for InputEvents.

The original field mInputFilter was replaced in this patch but not
removed. When a new InputFilter is added, we check that this unused
field mInputFilter != null before doing necessary teardown of the previous
filter. Since this is always null, this causes the previous filter to
not be disconnected when a new one is set with setInputFilter().

If the user toggles "Magnification Gestures" on and off twice in
Accessibility Settings, this will send the old and new InputFilters into a loop
sending and receiving touch events, locking up the device completely
until reboot.

Remove all references to the unused mInputFilter field.

Change-Id: Id28335d150a195af8747b4862deb897ae850d2cb