Torch: remind user flashlight is still on
Post a notification which informs the user the flashlight is still
enabled. But only do this when the user turns the screen off *while* the
flashlight is still on, so they see it next time they use the device.
Tapping on the notification will disable the flashlight.

Ported from PAC version.

Change-Id: Id5b19e615b686c915258e52ee4c22fceab62e315