Update patch set 4
Patch Set 4:

well that takes care of the listpreference being set to default of "off" anyway.  also you need to change the default value of the setting in the observer in frameworks side.  in this line:
https://github.com/SlimRoms/frameworks_base/blob/lp5.0/packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java#L1963
you need to change ONE_FINGER_QS_INTERCEPT_END (which has value of 1) to ONE_FINGER_QS_INTERCEPT_OFF (which has a value of 0).  then the behavior will default to off.

Patch-set: 4