Update patch set 11
Patch Set 11:

I noticed a weird behavior with the previous patchset, i was not able to drag the slider in the expanded volume panel while watching a youtube video. It never happened with older code, so i've restored it.

Atm, the only little issue is always the damned glitch when trying to drag the slider during a call :D

Patch-set: 11
Label: Code-Review=0