Update patch set 3
Patch Set 3:

if i may give my opinion...the reboot tile is supposed to give a quick access for the user to do a reboot without too many clicks and without using hw button (in this case power button). this is important for some users since their power button is broken.

the 'detail' view is introduced since lollipop. this view is usually used if the tile has more options. for reboot tile, the goal is only to reboot the device either normal or recovery. i find the old behavior (since kitkat) is absolutely brilliant.

concerning long press to power menu, wouldn't it be more faster if we create a power menu tile? as we are on this power-menu topic, wouldn't it be more better if we provide slimaction as a tile? so there won't be any questions like: could you make screenshot/power menu/screen off as a tile? we have this in slimkat, can't remember what it's called. i already ported this and called it as 'slimaction' tile. it uses the same dslv that has been used in navbar and slimpie to add/choose slimaction. in this tile, i made it like this: single press to switch between slimaction whereas long press to execute the actual slimaction without using detail view.

Patch-set: 3
Reviewer: Chandra Poerwanto <1000103@f8201d64-d729-4630-b91d-5f8be02869f0>
Label: Code-Review=0