Update patch set 8
Patch Set 8: -Code-Review

No, i made it as a notification led toggle  :P
Because i use to have the battery led still on but with a not disturbing color: 
http://s30.postimg.org/5r89btext/uploadfromtaptalk1433445617370.png
Other people also always disable it always because of an additional led on their charger, so a toggle for this led too would be a drawback.
If we wanna act on the battery led too, an option in the QS settings menu to enable/disable the battery led toggle for this tile would be the best choice i think. But i don't think i'm able to do that :D

Patch-set: 8