Update patch set 8
Patch Set 8:

I more detailed answer with theme engine is the fact the the resources themselves for FB and ANYWHERE besides settings to do even make it into the final apk. They are there when it is just named package.apk but in the process of finalizing there are somehow removed. Even if you decompile the apk and add the resources manually tho the only thing that changes is settings. I think it has to do with new theme engine permission setup as it seems to block many things but I could be wrong. Either way that's a better explanation of the actual issues faced when trying to use alongside theme engine, I guess could be a 5.1 issue as well since I never tried it on 5.0

Patch-set: 8
Label: Code-Review=0