[1/2] Slim Recent App Sidebar
One thing I was missing with SlimRecents was the possibility to launch your favorite apps from the recents screen. As SlimRecents covers only the half of the screen (more or less), there is enough space on the other edge to display an app sidebar, which I based on the ChameleonOS one.

Changes to the ChameleonOS sidebar include:
- No trigger areas, but show/hide with the SlimRecent screen
- Adapted look and animation to SlimRecent
- Longpress actions
- Slimer settings using ActionListViewSettings, also added possibility to configure sidebar size (roughly, as the size is adapted to the screen height in order to show full icons) and colors
- Currently no option to add folders to the sidebar, as this does not work with the current ActionListViewSettings without further work

Change-Id: I1f3edce2ed12297a7a7cb976f2a2315f86492192