Update patch set 5
Patch Set 5:

Not sure if that will solve the issue, there will be no rotation when doze is showing notifications as the accelerometer is sleeping.

The issue is if the device is in landscape when screen goes / is turned off, doze will be in landscape.

So we'd need a "set display to protrait when screen goes off", or "set display to protrait when doze is about to show".

Patch-set: 5
Label: Code-Review=0