MediaActionSound: Fix no capture sound when taking screenshot
When taking screenshot for the first time after device booted up,
no capture sound can be heard as the sample played while the sample
is still in loading progress.

Check the sample loading state when played, if play() failed due to
sample is not ready, then put the playback on completion of sample
loading.

Change-Id: Icfcfc4768d7d4cd59b4761905a9a1d479a300f30
CRs-Fixed: 958389