Update patch set 3
Patch Set 3:

This change does indeed provide the same number of tiles in the topmost default row as is specified in the settings (and displayed in other rows), but the horizontal spacing is greater than in the other rows, causing the tiles/text to be shifted off the edges a little.

Also, every tile in the default topmost row has a drop-down arrow beneath it, whether or not the tile has long-press or options functionality.

The spacing for a 4-across layout is OK.

Patch-set: 3